+++ Project/serverConfig.json
@@ -3,5 +3,3 @@
     "functionAppOneSignalTokenRegistrationUrl": "",
     "functionAppOneSignalTokenRegistrationApiKey":""
 }
-
-
